backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
supabase==2.13.0
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.27.2
aiosmtplib==3.0.1
yfinance==0.2.51
pandas==2.2.3
numpy==2.0.2
scipy==1.14.1
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==5.2.1
fastapi-cache2[redis]==0.2.2
PyJWT
